success! simulation runs TODO: rare sig errors at line 271, very rare. implement observer, clean up/dry code, add way to toggle DAS, Double only allowed after first 2 cards, stretch surrender togggle, start testing more stratsgit add . !

+++ BasicStrategy.h
@@ -12,13 +12,13 @@ class BasicStrategy {
         static const Action splitTable[10][10];
         
         static int getIndex(Rank dealerUpcard);
-        static Action shouldDeviatefromHard(int playerTotal, Rank dealerUpcard,int true_count);
-        static Action shouldDeviatefromSplit(Rank playerSplitRank, Rank dealerUpcard,int true_count);
+        static Action shouldDeviatefromHard(int playerTotal, Rank dealerUpcard,float true_count);
+        static Action shouldDeviatefromSplit(Rank playerSplitRank, Rank dealerUpcard,float true_count);
 
     public:
-        static Action getHardHandAction(int playerTotal, Rank dealerUpcard,int true_count);
+        static Action getHardHandAction(int playerTotal, Rank dealerUpcard,float true_count);
         static Action getSoftHandAction(int playerTotal, Rank dealerUpcard);
-        static Action getSplitAction(Rank playerSplitRank, Rank dealerUpcard,int true_count);
+        static Action getSplitAction(Rank playerSplitRank, Rank dealerUpcard,float true_count);
 
 };
